requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile setup.py
#
paho-mqtt==1.5.1
    # via ota (setup.py)
